add clear cache button to diag modal in Boilerplate (#10095)
